fix: address pr reviews
